+++ ledger-service/http-json/src/itpostgres/scala/http/HttpServicePostgresInt.scala
@@ -4,7 +4,9 @@
 package com.daml.http
 
 import com.codahale.metrics.MetricRegistry
-import com.daml.http.dbbackend.JdbcConfig
+import com.daml.dbutils
+import com.daml.dbutils.ConnectionPool
+import dbbackend.{DbStartupMode, JdbcConfig}
 import com.daml.metrics.Metrics
 import com.daml.testing.postgresql.PostgresAroundAll
 import org.scalatest.Inside
@@ -21,10 +23,24 @@ trait HttpServicePostgresInt extends AbstractHttpServiceIntegrationTestFuns with
   protected lazy val dao = dbbackend.ContractDao(jdbcConfig_)
 
   // has to be lazy because postgresFixture is NOT initialized yet
-  protected[this] def jdbcConfig_ = PostgresIntTest.defaultJdbcConfig(postgresDatabase.url)
+  protected[this] def jdbcConfig_ = HttpServicePostgresInt.defaultJdbcConfig(postgresDatabase.url)
 
   override protected def afterAll(): Unit = {
     dao.close()
     super.afterAll()
   }
+}
+
+object HttpServicePostgresInt {
+  def defaultJdbcConfig(url: => String) = JdbcConfig(
+    dbutils.JdbcConfig(
+      driver = "org.postgresql.Driver",
+      url = url,
+      user = "test",
+      password = "",
+      tablePrefix = "some_nice_prefix_",
+      poolSize = ConnectionPool.PoolSize.Integration,
+    ),
+    dbStartupMode = DbStartupMode.CreateOnly,
+  )
 }
